+++ packages/algob/test/stubs/algo-operator.ts
@@ -17,10 +17,6 @@ export class AlgoOperatorDryRunImpl implements AlgoOperator {
     return mockAlgod;
   };
 
-  getDelegatedLsig (lsig: string): Object | undefined {
-    throw new Error("Not implemented");
-  }
-
   getAssetByID (assetIndex: number | bigint): Promise<modelsv2.Asset> {
     return new Promise((resolve, reject) => {
       assetIndex === 1n ? resolve(mockAssetInfo) : reject(new Error("Not implemented"));